started spelling suggestions

+++ internal/lib.go
@@ -5,8 +5,17 @@ import (
 	"path/filepath"
 )
 
+func DoesPathExists(p string) bool {
+	if _, err := os.Stat(p); os.IsNotExist(err) {
+		return false
+	}
+	return true
+}
+
 func GetDefaultFontPath() string {
 	fontPath := filepath.Join(os.TempDir(), "picker_font.ttf")
-	os.WriteFile(fontPath, DefaultFont, 0777)
+	if !DoesPathExists(fontPath) {
+		os.WriteFile(fontPath, DefaultFont, 0777)
+	}
 	return fontPath
 }
